Vital Sign
I am in the midst of restructuring and this commit is to show what I've been doing since the last commit.

I have:

Decoupled the window from the renderer, renamed some variables, begun working on the UI by decoupling it from the renderer, added license information and updated README to include more information and rewrote some parts of it.